Rename dummy Link and use real Link instead of dummy
